Moved remaining road classes to Roads namespace
