+++ src/OpenSage.Game/Content/AssetStore.cs
@@ -21,6 +21,7 @@ using OpenSage.Logic.Object;
 using OpenSage.Logic.Object.Damage;
 using OpenSage.Logic.Pathfinding;
 using OpenSage.Terrain;
+using OpenSage.Terrain.Roads;
 using Veldrid;
 
 namespace OpenSage.Content
